requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
authlib>=1.3.0
httpx>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.0
click>=8.0.0
orjson>=3.8.0
python-dotenv>=0.19.0
tqdm>=4.65.0
pathlib>=1.0.1
regex>=2023.0.0
duckdb>=1.3.1
pandas>=2.0.0
posthog>=6.0.0
itsdangerous>=2.1.0
